Update README.md
# My-Profile-Page
This is my Mini project assessment for my first 3MTT Project: Personal profile Page (HTML only).
In this project, I created my first profile page using semantic html elements.
the page contains element such as header which include the title of the page, an image of myself (placeholder image) as well as a navigation bar.
The body of the web page contains different sections which includes About me, Hobbies and interest, Projects and Contact me.
The about me sections gives an introduction of my bio which is well aligned using the <p>paragraph</p> element as well as the <ul>unordered list</ul> elements. Other elements of html such as <strong>strong</strong> was used to create emphasis.
Same element applied to the About me section was also utilized in the other sections of the web page.
The webpage also contains a footer section as directed by the instructor. this is aimed at meeting all the requirement of the project.
The html file is hereby attached in a well-organized and readable fashion.
Thank you!